requirements.txt
torch == 1.10.0
torchvision == 0.11.0
numpy == 1.20.2
MinkowskiEngine == 0.5.3
nuscenes-devkit == 1.1.9
mmcv-full==1.4.0
copy
tqdm
setuptools==59.5.0
pytorch_lightning == 1.4.0
Pillow == 8.3.1
pyquaternion
multiprocessing
scikit-image == 0.18.2
spconv == 1.2.1
torchmetrics == 0.4.0